+++ views/auth/login.hbs
@@ -13,7 +13,7 @@
 
     <button class="btn btn-success">Log In</button>
 
-    <p>If you don't have an account yet<a href="/auth/signup" class="link-success link-offset-3 link-underline-opacity-100">Signup</a></p>
+    <p>If you don't have an account yet <a href="/auth/signup" class="link-success link-offset-3 link-underline-opacity-100">Signup</a></p>
 
     {{#if errorMessage}}
       <div class="alert alert-warning" role="alert">
