+++ views/error.hbs
@@ -1,2 +1,3 @@
 <h1>Error!</h1>
 <h2>There was an error on the server! Check the console!</h2>
+<h3>{{errorMessage}}</h3>
