+++ views/layout.hbs
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
     <title>Topgamesports</title>
-    <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
+    <link rel="shortcut icon" href="favicon.ico" type="image/x-icon" />
     <link
       href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/css/bootstrap.min.css"
       rel="stylesheet"
@@ -39,14 +39,14 @@
         >
           <span class="navbar-toggler-icon"></span>
         </button>
-        <div class="collapse navbar-collapse" id="navbarScroll" >
+        <div class="collapse navbar-collapse" id="navbarScroll">
           <ul
             class="navbar-nav me-auto my-2 my-lg-0 navbar-nav-scroll"
             style="--bs-scroll-height: 100px;"
           >
             {{#if isUserActive}}
               <li class="nav-item">
-                <a class="nav-link active" aria-current="page" href="/game" ><svg
+                <a class="nav-link active" aria-current="page" href="/game"><svg
                     width="20"
                     height="20"
                     viewBox="0 0 48 48"
@@ -83,6 +83,24 @@
                     />
                   </svg>
                   Events</a>
+              </li>
+              <li class="nav-item">
+                <a
+                  class="nav-link active"
+                  aria-current="page"
+                  href="/event/esports"
+                ><svg
+                    width="20"
+                    height="20"
+                    viewBox="0 0 24 24"
+                    xmlns="http://www.w3.org/2000/svg"
+                  >
+                    <path
+                      fill="#000000"
+                      d="M9.103.435a1.753 1.753 0 0 1 1.652-.362c.218.072.406.203.609.333c.16.101.348.145.493.261c.072.058.014.16.014.232c.189.45.29.942.13 1.42c-.13.16-.362.203-.55.276c-.03.203.043.39.072.594c-.043.029-.072.058-.116.087c.276-.015.551-.073.827-.13c.101-.102.26-.044.39-.059c.015-.203.088-.391.088-.594a.34.34 0 0 0 .116-.029c.014.145 0 .29.043.435c.073.058.189.029.276.043c0 .058 0 .116.014.174c1.681-.014 3.377 0 5.058 0v.247h.16V2.42h.188c0 .26-.014.507 0 .768a.11.11 0 0 0 .073.029c0 .029.014.087.014.116c.058-.058.13-.102.218-.073c.014.044.029.087.043.145c-.058.058-.087.13-.058.218c.464.014.928 0 1.406 0c.044-.058.116-.087.189-.116a.464.464 0 0 1 .087.058h.68a.874.874 0 0 1 .015.333h-.696a.46.46 0 0 1-.087.073c-.072-.044-.13-.073-.202-.116h-1.261c-.247.072-.508.058-.754.014v.275H16.16a.532.532 0 0 1-.29.13c.044.218-.202.276-.29.436a.45.45 0 0 1-.231.174c-.073.449.087.87.203 1.29c-.13.029-.276.072-.406.101c-.073.29-.145.594-.203.884a.848.848 0 0 1-.507.58c-.174.203-.406.406-.682.42c-.101.03-.174-.043-.246-.101c-.362.029-.696-.145-1.015-.29c-.347-.16-.666-.363-1-.55c.03.202-.072.376-.145.55c.16.072.377.13.464.304c.058.13.116.276.116.435c-.014.522-.072 1.044-.101 1.565c.014.377-.174.754-.435 1.03c-.174-.015-.319-.088-.478-.16c-.058.13-.189.26-.116.42c.058.189.058.392.145.566a6.674 6.674 0 0 1 1.217 1.753c.304.624.536 1.276.783 1.928c.043 0 .13-.015.174-.015c.058.189 0 .406.116.58c.101.16.072.348.072.536c-.029.435-.058.87-.101 1.305c-.03.304-.102.609-.145.913c.014.232.116.464.101.696c-.014.217-.014.449-.188.608c.014.493-.116.986.058 1.464c.232.32.493.623.768.899c.304.145.667.174.928.435c.1.16.043.347.014.521a6.732 6.732 0 0 1-1.87 0c-.246-.058-.478-.159-.724-.188c-.334.014-.725.145-1.03-.087c-.028-.391.117-.768.189-1.145c.029-.13.16-.217.145-.348c-.03-.45-.058-.913-.087-1.362c-.058-.03-.16-.058-.145-.145c0-.218-.072-.435-.13-.638a10.821 10.821 0 0 1-.16-1.681c-.014-.16.087-.276.203-.377c.03-.246.058-.507.073-.754c-.044-.13-.145-.232-.203-.348c-.261.03-.638.087-.797-.188c-.377-.565-.769-1.145-1.145-1.71c-.16-.015-.348 0-.493-.102c-.16-.174-.261-.405-.363-.623c-.043.174-.072.362-.174.507c-.087.145-.231.247-.333.377c-.101.232-.188.464-.275.696c-.102.29-.247.58-.276.899a1.543 1.543 0 0 1-.101.449c-.073.116-.203.16-.319.217c-.087.189-.145.377-.275.537c-.087.101-.232.13-.29.26c-.058.174-.145.334-.218.493c-.029.174.087.363.03.536c-.102.435-.32.841-.522 1.232c-.102.29-.174.595-.32.87c-.057.116-.202.145-.318.16c-.13.318-.276.623-.362.956a5.437 5.437 0 0 0-.03.971c0 .145.088.261.146.391c.057.174.014.348-.015.522c-.565.073-1.145.13-1.696-.043c-.058-.044-.043-.116-.058-.174c-.043-.261-.072-.536.015-.783c.188-.681.348-1.362.536-2.043c-.072-.073-.174-.13-.174-.247c-.014-.188 0-.391.044-.58c.087-.319.318-.565.434-.87c.044-.13.03-.26.044-.39c0-.305.174-.551.304-.812c.13-.218.232-.45.406-.638c.116-.101.116-.26.203-.391c.087-.16.232-.29.232-.479c.029-.231-.058-.463-.03-.695c.059-.681.19-1.348.305-2.03c-.058-.072-.145-.144-.174-.246c.015-.072.03-.13.044-.203l-.13-.217c.057-.087.115-.188.173-.275c-.058-.044-.13-.102-.188-.145c.072-.218.043-.537.304-.638c.03.014.102.029.13.043c-.043-.376-.043-.768-.086-1.145a3.534 3.534 0 0 1-.073-1.232c.145-.217.42-.304.667-.318c-.319-.073-.638-.102-.942-.203c-.015-.261.029-.522.072-.783c.13-.507.073-1.029.145-1.55c.073-.146.261-.175.42-.146c.189.03.377-.029.566-.087c0-.072.014-.16 0-.232c-.116-.507-.087-1.029 0-1.521c.116-.638.377-1.261.855-1.71c.319-.305.783-.45 1.217-.435c.145 0 .232.145.348.232c.058-.058.116-.116.16-.174c-.073-.247-.189-.508-.174-.769c.029-.58.217-1.174.652-1.565m4.522 4.102c.029.029.029.029 0 0m.203.029c.014.101.043.203-.015.29c-.072.029-.16.029-.232.072c.203 0 .406.015.61 0c.159-.043.1-.246.1-.362c-.1-.116-.318-.044-.463 0m-.507.609c.145.159.217.405.347.565c.189-.247.334-.508.551-.725c-.26-.015-.522.015-.782-.015c-.044.058-.073.116-.116.174Z"
+                    />
+                  </svg>
+                  Esports</a>
               </li>
             {{/if}}
             {{#if isAdminActive}}
